// src/app/shared-elephant-bet/models/bilan-jeu.ts
export class BilanJeu {
	id: number;
	dateCreation: string;
	nomJeu: string;
	nombreJeu: number;
	volumeJeu: number;
	nombreGains: number;
	volumeGains: number;
	balance: number;
	nombreGainsSuperieurLimite: number;
	nombreGainsInferieurLimite: number;
	volumeGainsInferieurLimite: number;
	volumeGainsSuperieurLimite: number;
}